Create RedirectCreatePageButton component
PS.: reused from CreateUserButton component
